+++ CHANGELOG.rst
@@ -14,10 +14,20 @@ Change Log
 Unreleased
 ----------
 
-[0.60.1] - 2018-01-08
+[0.61.2] - 2018-01-09
+---------------------
+
+* Add missing migrations for sap_success_factors and degreed.
+
+[0.61.1] - 2018-01-09
 ---------------------
 
 * Update django admin list view for enterprise customer model.
+
+[0.61.0] - 2018-01-09
+---------------------
+
+* SuccessFactors Admin Update: Enterprise Customer Configuration.
 
 [0.60.0] - 2018-01-03
 ---------------------